requirements.txt
alembic==1.10.2
anyio==3.6.2
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.0.1
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
cryptography==40.0.1
debugpy==1.6.6
decorator==5.1.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
executing==1.2.0
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
ipykernel==6.21.1
ipython==8.9.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.2
jupyter_core==5.2.0
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
orjson==3.8.8
packaging==23.0
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
platformdirs==3.0.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.5
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pywin32==305
PyYAML==6.0
pyzmq==25.0.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.0
stack-data==0.6.2
starlette==0.26.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.21.1
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
wincertstore==0.2
